Restructure the neuro-integration file structure
